refactor DynamicCategoryIcon
